费用设置

+++ BtsSys/src/com/scttsc/common/util/StoreUtil.java
@@ -1,17 +1,16 @@
 package com.scttsc.common.util;
 
-import com.scttsc.business.util.Constants;
-import org.apache.commons.io.FileUtils;
-import org.apache.log4j.Logger;
-
 import java.io.File;
 import java.io.IOException;
+
+import org.apache.commons.io.FileUtils;
+import org.apache.log4j.Logger;
 
 /**
  * Created by _think on 2014/11/16.
  */
 public class StoreUtil {
-    Logger LOG = Logger.getLogger(StoreUtil.class);
+    private static Logger LOG = Logger.getLogger(StoreUtil.class);
 
     static {
         init();
@@ -41,16 +40,35 @@ public class StoreUtil {
     public static void copyFile(String uuid,String descPath) {
         File tmpFile = new File(storeTmpDir() + "/" + uuid);
         try {
-            if (tmpFile != null) {
+            if (tmpFile != null && tmpFile.exists()) {
                 File formalFile = new File(descPath+ "/" + uuid);
                 FileUtils.copyFile(tmpFile, formalFile);
-                if(tmpFile.exists()){
-                    tmpFile.delete();
-                }
+                tmpFile.delete();
             }
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+    
+    public static void deleteFile(String filePath) throws Exception{
+    	try {
+    		File file = new File(filePath);
+    		if(file.exists()&&file.isFile()){
+    			file.delete();
+    		}
+		} catch (Exception e) {
+			LOG.error(e.getMessage(), e);
+		}
+    }
+    
+    public static void deleteDir(String dirPath) throws Exception{
+    	File file = new File(dirPath);
+    	if(file.isDirectory()){
+    		for(File file2 : file.listFiles()){
+    			if(file2.exists()&&file2.isFile())
+    				file2.delete();
+    		}
+    	}
     }
 
     /**
